+++ mesos/mesos.go
@@ -12,7 +12,7 @@ import (
 	logrus "github.com/AVENTER-UG/mesos-m3s/logger"
 	mesosproto "github.com/AVENTER-UG/mesos-m3s/proto"
 	cfg "github.com/AVENTER-UG/mesos-m3s/types"
-	"github.com/gogo/protobuf/jsonpb"
+	"google.golang.org/protobuf/encoding/protojson"
 )
 
 // Mesos include all the current vars and global config
@@ -22,10 +22,10 @@ type Mesos struct {
 }
 
 // Marshaler to serialize Protobuf Message to JSON
-var marshaller = jsonpb.Marshaler{
-	EnumsAsInts: false,
-	Indent:      " ",
-	OrigName:    true,
+var marshaller = protojson.MarshalOptions{
+	UseEnumNumbers: false,
+	Indent:         " ",
+	UseProtoNames:  true,
 }
 
 // New will create a new API object
@@ -42,7 +42,7 @@ func New(cfg *cfg.Config, frm *cfg.FrameworkConfig) *Mesos {
 func (e *Mesos) Revive() {
 	logrus.WithField("func", "mesos.Revive").Debug("Revive Tasks")
 	revive := &mesosproto.Call{
-		Type: mesosproto.Call_REVIVE,
+		Type: mesosproto.Call_REVIVE.Enum(),
 	}
 	err := e.Call(revive)
 	if err != nil {
@@ -54,7 +54,7 @@ func (e *Mesos) Revive() {
 func (e *Mesos) SuppressFramework() {
 	logrus.WithField("func", "mesos.SuppressFramework").Info("Framework Suppress")
 	suppress := &mesosproto.Call{
-		Type: mesosproto.Call_SUPPRESS,
+		Type: mesosproto.Call_SUPPRESS.Enum(),
 	}
 	err := e.Call(suppress)
 	if err != nil {
@@ -67,13 +67,13 @@ func (e *Mesos) Kill(taskID string, agentID string) error {
 	logrus.WithField("func", "mesos.Kill").Info("Kill task ", taskID)
 	// tell mesos to shutdonw the given task
 	err := e.Call(&mesosproto.Call{
-		Type: mesosproto.Call_KILL,
+		Type: mesosproto.Call_KILL.Enum(),
 		Kill: &mesosproto.Call_Kill{
-			TaskID: mesosproto.TaskID{
-				Value: taskID,
+			TaskId: &mesosproto.TaskID{
+				Value: &taskID,
 			},
-			AgentID: &mesosproto.AgentID{
-				Value: agentID,
+			AgentId: &mesosproto.AgentID{
+				Value: &agentID,
 			},
 		},
 	})
@@ -83,8 +83,8 @@ func (e *Mesos) Kill(taskID string, agentID string) error {
 
 // Call will send messages to mesos
 func (e *Mesos) Call(message *mesosproto.Call) error {
-	message.FrameworkID = e.Framework.FrameworkInfo.ID
-	body, err := marshaller.MarshalToString(message)
+	message.FrameworkId = e.Framework.FrameworkInfo.Id
+	body, err := marshaller.Marshal(message)
 
 	if err != nil {
 		logrus.WithField("func", "mesos.Call").Debug("Could not Marshal message:", err.Error())
@@ -119,38 +119,38 @@ func (e *Mesos) Call(message *mesosproto.Call) error {
 		body, err := io.ReadAll(res.Body)
 		if err != nil {
 			logrus.WithField("func", "mesos.Call").Error("Call Handling (could not read res.Body)")
-			return fmt.Errorf("Error %d", res.StatusCode)
+			return fmt.Errorf("error %d", res.StatusCode)
 		}
 
 		logrus.WithField("func", "mesos.Call").Error("Call Handling: ", string(body))
-		return fmt.Errorf("Error %d", res.StatusCode)
+		return fmt.Errorf("error %d", res.StatusCode)
 	}
 
 	return nil
 }
 
 // DecodeTask will decode the key into an mesos command struct
-func (e *Mesos) DecodeTask(key string) cfg.Command {
-	var task cfg.Command
+func (e *Mesos) DecodeTask(key string) *cfg.Command {
+	var task *cfg.Command
 	if key != "" {
 		err := json.NewDecoder(strings.NewReader(key)).Decode(&task)
 		if err != nil {
 			logrus.WithField("func", "scheduler.DecodeTask").Error("Could not decode task: ", err.Error())
-			return cfg.Command{}
+			return &cfg.Command{}
 		}
 		return task
 	}
-	return cfg.Command{}
+	return &cfg.Command{}
 }
 
 // GetOffer get out the offer for the mesos task
-func (e *Mesos) GetOffer(offers *mesosproto.Event_Offers, cmd cfg.Command) (mesosproto.Offer, []mesosproto.OfferID) {
-	var offerIds []mesosproto.OfferID
-	var offerret mesosproto.Offer
+func (e *Mesos) GetOffer(offers *mesosproto.Event_Offers, cmd *cfg.Command) (*mesosproto.Offer, []*mesosproto.OfferID) {
+	var offerIds []*mesosproto.OfferID
+	var offerret *mesosproto.Offer
 
 	for n, offer := range offers.Offers {
 		logrus.WithField("func", "mesos.GetOffer").Debug("Got Offer From:", offer.GetHostname())
-		offerIds = append(offerIds, offer.ID)
+		offerIds = append(offerIds, offer.Id)
 
 		if cmd.TaskName == "" {
 			continue
@@ -164,16 +164,16 @@ func (e *Mesos) GetOffer(offers *mesosproto.Event_Offers, cmd cfg.Command) (meso
 			continue
 		}
 		offerret = offers.Offers[n]
-		offerIds = e.removeOffer(offerIds, offerret.ID.Value)
+		offerIds = e.removeOffer(offerIds, offerret.Id.GetValue())
 	}
 	return offerret, offerIds
 }
 
 // remove the offer we took from the list
-func (e *Mesos) removeOffer(offers []mesosproto.OfferID, clean string) []mesosproto.OfferID {
-	var offerIds []mesosproto.OfferID
+func (e *Mesos) removeOffer(offers []*mesosproto.OfferID, clean string) []*mesosproto.OfferID {
+	var offerIds []*mesosproto.OfferID
 	for _, offer := range offers {
-		if offer.Value != clean {
+		if offer.GetValue() != clean {
 			offerIds = append(offerIds, offer)
 		}
 	}
@@ -181,15 +181,15 @@ func (e *Mesos) removeOffer(offers []mesosproto.OfferID, clean string) []mesospr
 }
 
 // DeclineOffer will decline the given offers
-func (e *Mesos) DeclineOffer(offerIds []mesosproto.OfferID) *mesosproto.Call {
+func (e *Mesos) DeclineOffer(offerIds []*mesosproto.OfferID) *mesosproto.Call {
 
 	logrus.WithField("func", "scheduler.HandleOffers").Debug("Offer Decline: ", offerIds)
 
 	refuseSeconds := 120.0
 
 	decline := &mesosproto.Call{
-		Type: mesosproto.Call_DECLINE,
-		Decline: &mesosproto.Call_Decline{OfferIDs: offerIds, Filters: &mesosproto.Filters{
+		Type: mesosproto.Call_DECLINE.Enum(),
+		Decline: &mesosproto.Call_Decline{OfferIds: offerIds, Filters: &mesosproto.Filters{
 			RefuseSeconds: &refuseSeconds,
 		},
 		},
@@ -199,7 +199,7 @@ func (e *Mesos) DeclineOffer(offerIds []mesosproto.OfferID) *mesosproto.Call {
 
 // IsRessourceMatched - check if the ressources of the offer are matching the needs of the cmd
 // nolint:gocyclo
-func (e *Mesos) IsRessourceMatched(ressource []mesosproto.Resource, cmd cfg.Command) bool {
+func (e *Mesos) IsRessourceMatched(ressource []*mesosproto.Resource, cmd *cfg.Command) bool {
 	mem := false
 	cpu := false
 	ports := true
@@ -217,8 +217,10 @@ func (e *Mesos) IsRessourceMatched(ressource []mesosproto.Resource, cmd cfg.Comm
 			if v.GetName() == "ports" {
 				for _, taskPort := range cmd.DockerPortMappings {
 					for _, portRange := range v.GetRanges().Range {
-						if taskPort.HostPort >= uint32(portRange.Begin) && taskPort.HostPort <= uint32(portRange.End) {
-							logrus.WithField("func", "mesos.IsRessourceMatched").Debug("Matched Offer TaskPort: ", taskPort.HostPort)
+						portBegin := uint32(portRange.GetBegin())
+						portEnd := uint32(portRange.GetEnd())
+						if *taskPort.HostPort >= portBegin && *taskPort.HostPort <= portEnd {
+							logrus.WithField("func", "mesos.IsRessourceMatched").Debug("Matched Offer TaskPort: ", taskPort.GetHostPort())
 							logrus.WithField("func", "mesos.IsRessourceMatched").Debug("Matched Offer RangePort: ", portRange)
 							ports = ports || true
 							break
@@ -284,19 +286,19 @@ func (e *Mesos) GetAgentInfo(agentID string) cfg.MesosSlaves {
 }
 
 // GetNetworkInfo get network info of task
-func (e *Mesos) GetNetworkInfo(taskID string) []mesosproto.NetworkInfo {
+func (e *Mesos) GetNetworkInfo(taskID string) []*mesosproto.NetworkInfo {
 	task := e.GetTaskInfo(taskID)
 
 	if len(task.Tasks) > 0 {
 		for _, status := range task.Tasks[0].Statuses {
 			if status.State == "TASK_RUNNING" {
-				var netw []mesosproto.NetworkInfo
+				var netw []*mesosproto.NetworkInfo
 				netw = append(netw, status.ContainerStatus.NetworkInfos[0])
 				return netw
 			}
 		}
 	}
-	return []mesosproto.NetworkInfo{}
+	return []*mesosproto.NetworkInfo{}
 }
 
 // GetTaskInfo get the task object to the given ID
@@ -311,7 +313,7 @@ func (e *Mesos) GetTaskInfo(taskID string) cfg.MesosTasks {
 	if !e.Framework.MesosSSL {
 		protocol = "http"
 	}
-	req, _ := http.NewRequest("POST", protocol+"://"+e.Framework.MesosMasterServer+"/tasks/?task_id="+taskID+"&framework_id="+e.Framework.FrameworkInfo.ID.GetValue(), nil)
+	req, _ := http.NewRequest("POST", protocol+"://"+e.Framework.MesosMasterServer+"/tasks/?task_id="+taskID+"&framework_id="+e.Framework.FrameworkInfo.Id.GetValue(), nil)
 	req.Close = true
 	req.SetBasicAuth(e.Framework.Username, e.Framework.Password)
 	req.Header.Set("Content-Type", "application/json")