+++ scheduler/heatbeat.go
@@ -98,7 +98,7 @@ func (e *Scheduler) CheckState() {
 			task.Discovery = e.changeDiscoveryInfo(task)
 
 			// add task to communication channel
-			e.Framework.CommandChan <- task
+			e.Framework.CommandChan <- *task
 
 			e.Redis.SaveTaskRedis(task)