+++ changelog.md
@@ -33,6 +33,7 @@
 - Fix and optimise mesos offerhandling 
 - Add support for a TCP port beside HTTP and HTTPS at the Kubernetes agent. It will be configures and
   enabled with the env variable `K3S_AGENT_TCP_PORT`.
+- Migrate from gogo protobuf to golang protobuf. Update mesos proto files.
   
 
 ## v0.4.1